seeder and db fix

+++ IdleHabits-Backend/database/migrations/2025_05_10_193057_create_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedMediumInteger('base_price');
             $table->unsignedInteger('base_frequency'); // in milliseconds, how often the item will yield gold
             $table->unsignedInteger('base_gold'); // how much gold the item will yield
-            $table->unsignedSmallInteger('price_multiplier')->default(1); // how much the price of the item increases with each purchase
+            $table->double('price_multiplier')->default(1); // how much the price of the item increases with each purchase
             $table->timestamps();
         });
     }
